Move types into src
