+++ src/components/GifPreview.tsx
@@ -2,7 +2,7 @@
 "use client";
 
 import { useState } from "react";
-import { Gif } from "../../types/Gif";
+import { Gif } from "../types/Gif";
 import { Clipboard } from "./Clipboard";
 
 export const GifPreview = ({
