+++ src/components/MoreGifs.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { startTransition, useActionState, useEffect, useState } from "react";
-import { Gif, GifsResult } from "../../types/Gif";
+import { Gif, GifsResult } from "../types/Gif";
 import { useIntersectionObserver } from "../hooks/useIntersectionObserver";
 import { handleSearchGifs } from "../actions";
 import { IconSpinner } from "../icons/IconSpinner";
