+++ src/types/Gif.ts
